Implementation of graceperiod endpoint.
Please, see https://gitlab.com/supercash/services/parking-lot-service/-/wikis/API/5.-Payment for more info.

+++ src/main/java/cash/super_/platform/service/parkingplus/autoconfig/ParkingPlusProperties.java
@@ -64,7 +64,7 @@ public class ParkingPlusProperties extends SupercashServiceProperties {
   private String saleNameStartWith = "SUPERCASH";
 
   @NotNull
-  private long gracePeriod;
+  private Integer gracePeriod;
 
   public String getUserKey() {
     return userKey;
@@ -190,11 +190,11 @@ public class ParkingPlusProperties extends SupercashServiceProperties {
     this.saleNameStartWith = saleNameStartWith;
   }
 
-  public long getGracePeriod() {
+  public Integer getGracePeriod() {
     return gracePeriod;
   }
 
-  public void setGracePeriod(long gracePeriod) {
+  public void setGracePeriod(Integer gracePeriod) {
     this.gracePeriod = gracePeriod;
   }
